requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
diskcache==5.6.3
filelock==3.19.1
fsspec==2025.9.0
hf-xet==1.1.10
hf_transfer==0.1.9
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
sympy==1.14.0
torch==2.8.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
